--- v2/nacos/common/grpc/connection_event_listener.py
@@ -1,10 +1,0 @@
-from abc import ABC, abstractmethod
-class IConnectionEventListener(ABC):
-
-  @abstractmethod
-  def on_connected(self) -> None:
-      pass
-
-  @abstractmethod
-  def on_disconnect(self) -> None:
-      pass
