+++ v2/nacos/transport/connection.py
@@ -1,13 +1,8 @@
 from abc import ABC, abstractmethod
 from grpc import Channel
-from .remote.rpc_request import IRequest
-from .remote.rpc_response import IResponse
-
-class RpcClient:
-    pass
-
-class ServerInfo:
-    pass
+from  v2.nacos.common.model.request import IRequest
+from  v2.nacos.common.model.response import IResponse
+from  v2.nacos.transport.rpc_client import RpcClient
 
 class IConnection(ABC):
     @abstractmethod
@@ -37,18 +32,17 @@ class IConnection(ABC):
 
 class Connection(IConnection):
     def __init__(self, conn: Channel, connection_id: str, server_info: ServerInfo):
-        self._conn = conn  # This should be an actual gRPC connection object 
+        self._conn = conn  
         self._connection_id = connection_id
         self._abandon = False
         self._server_info = server_info
 
     def request(self, request: IRequest, timeout_mills: int, client: RpcClient) -> IResponse:
-        # Implement the request logic here
-        pass #考虑去掉这个无用的定义
+        pass 
 
     def close(self):
         if self._conn:
-            self._conn.close()  # Actual gRPC connection close logic
+            self._conn.close()  
 
     def get_connection_id(self) -> str:
         return self._connection_id
