+++ v2/nacos/transport/rpc_client_factory.py
@@ -1,0 +1,4 @@
+class ClientMap:
+
+  def __init__(self):
+      self.client_map = {}
